Separate cli_to_hub_listener.go to smaller files.
We're seeing if different files per service command will help in the
long run.
